refactor: sentry util
